editGameJS

+++ src/Event/startGameEvent.js
@@ -15,15 +15,18 @@ export default class StartGameEvent extends Event{
     function* gen() {
       if (Game.debug) Game.scene.ChangeState(STATE.STAGE);
       else Game.scene.ChangeState(STATE.LOADING);
-      //LOADING ANIMATION
-      MapData.DeleteStage();
-      for (let i = 0; i < 150; i++) {
+      if(!Game.debug){
+        //LOADING ANIMATION
+        MapData.DeleteStage();
+        for (let i = 0; i < 150; i++) {
           yield;
+        }
+        //LOADING終了
+        UIManager.Clean();
+        Game.scene.ChangeState(STATE.TITLE);
+        Audio.PlayBGM("title", 0);
       }
-      //LOADING終了
-      UIManager.Clean();
-      Game.scene.ChangeState(STATE.TITLE);
-      Audio.PlayBGM("title", 0);
+
       if (Game.debug) MapData.CreateStage(Game.stage, "ENTER");
       else MapData.CreateStage(0, "ENTER");
